Added refactored dashboard page
